+++ lib/src/mobile_scanner_controller.dart
@@ -175,20 +175,19 @@ class MobileScannerController extends ValueNotifier<MobileScannerState> {
 
   void _throwIfNotInitialized() {
     if (!value.isInitialized) {
-      throw const MobileScannerException(
+      throw MobileScannerException(
         errorCode: MobileScannerErrorCode.controllerUninitialized,
         errorDetails: MobileScannerErrorDetails(
-          message: 'The MobileScannerController has not been initialized.',
+          message: MobileScannerErrorCode.controllerUninitialized.message,
         ),
       );
     }
 
     if (_isDisposed) {
-      throw const MobileScannerException(
+      throw MobileScannerException(
         errorCode: MobileScannerErrorCode.controllerDisposed,
         errorDetails: MobileScannerErrorDetails(
-          message:
-              'The MobileScannerController was used after it has been disposed.',
+          message: MobileScannerErrorCode.controllerDisposed.message,
         ),
       );
     }
@@ -298,11 +297,10 @@ class MobileScannerController extends ValueNotifier<MobileScannerState> {
   /// If the permission is denied on iOS, MacOS or Web, there is no way to request it again.
   Future<void> start({CameraFacing? cameraDirection}) async {
     if (_isDisposed) {
-      throw const MobileScannerException(
+      throw MobileScannerException(
         errorCode: MobileScannerErrorCode.controllerDisposed,
         errorDetails: MobileScannerErrorDetails(
-          message:
-              'The MobileScannerController was used after it has been disposed.',
+          message: MobileScannerErrorCode.controllerDisposed.message,
         ),
       );
     }
@@ -354,13 +352,6 @@ class MobileScannerController extends ValueNotifier<MobileScannerState> {
         );
       }
     } on MobileScannerException catch (error) {
-      // If the controller is already initialized, ignore the error.
-      // Starting the controller while it is already started, or in the process of starting, is redundant.
-      if (error.errorCode ==
-          MobileScannerErrorCode.controllerAlreadyInitialized) {
-        return;
-      }
-
       // The initialization finished with an error.
       // To avoid stale values, reset the camera direction,
       // output size, torch state and zoom scale to the defaults.