+++ lib/src/web/mobile_scanner_web.dart
@@ -277,10 +277,10 @@ class MobileScannerWeb extends MobileScannerPlatform {
         );
       }
 
-      throw const MobileScannerException(
+      throw MobileScannerException(
         errorCode: MobileScannerErrorCode.controllerAlreadyInitialized,
         errorDetails: MobileScannerErrorDetails(
-          message: 'The scanner was already started.',
+          message: MobileScannerErrorCode.controllerAlreadyInitialized.message,
         ),
       );
     }